Add JDK-only implementation ConcurrentMapApplicationErrorDao
* Add ConcurrentMapApplicationErrorDao, an implementation of
  ApplicationErrorDao which has no dependencies outside the JDK. This
  implementation uses a ConcurrentMap to store ApplicationErrors
* Change error message in Jdbi3ApplicationErrorDao#incrementCount to
  be consistent with ConcurrentMapApplicationErrorDao
* Use KiwiPreconditions#checkArgumentIsNull to check that the
  ApplicationError has a null ID in Jdbi3ApplicationErrorDao#insertError
* Refactor tests so that a new AbstractApplicationErrorDaoTest is the
  abstract base test class for the JDBI and ConcurrentMap DAO
  implementations, and a new AbstractJdbi3ApplicationErrorDaoTest
  is now the parent of the H2 and Postgres test classes. All the actual
  tests are defined in AbstractApplicationErrorDaoTest and the
  subclasses define the specific implementation to be tested, along
  with any necessary setup.

Misc:

* Fix minor javadoc errors in ApplicationError

Closes #208
